(PUP-1916) Allow cert clean when node has a CSR
Prior to this commit, running `puppet cert clean` against a
certname that does not have a signed cert but has a CSR would
cause a failure because we tried to revoke the cert when
of course you cannot revoke an unsigned cert.

After this commit, we check to make sure the certname passed into
puppet cert clean is not currently waiting to be signed.  If the
cert is waiting to be signed then we do not try to revoke it and
move along to destroying which clears the CSR.

If you pass a certname that has no signed cert and no CSR then
cert clean will fail in the same way that it did before.