(PUP-1916) Correctly revoke mulitple hosts when one has a CSR
Prior to this commit, if you passed multiple hosts to
puppet cert clean and one of the hosts had a CSR pending then
none of the hosts would get their certs revoked but they
would be destroyed.

After this commit, we remove the hosts with a CSR from
the list of hosts that should get their certs revoked.