(MAINT) Add the jobname and PR number to the cache key
Since we don't have a Gemfile.lock we use the jobname and
PR number to get something specific enough.

When a new PR is opened it will restore a cache from a previous
PR and save a new cache.  This means that as new versions of Gems
come out we will continue to save off newer caches with those
version in it.  If we don't include the PR number then we could
get a longstanding cache that needs new versions of Gems installed
on top of it.